+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Merge cy_borg-core content module into cy_borg system.
 - Add in-game README as a journal entry compendium.
+- Add "Make NPC" button.
+- Add Roles random table.
+- Tweak various random description tables to be third-person.
 
 # 1.1.0